3-web-interface/requirements.txt
flask>=2.3.0
ultralytics>=8.0.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
Pillow>=9.0.0
